+++ src/containers/App.js
@@ -1,10 +1,15 @@
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
+import { withRouter } from 'react-router-dom';
 
 import App from '../components/App';
 import { askForGeoLocation } from '../ducks/geoLocation';
+import { getConnectedUserToken } from '../ducks/connectedUser';
 
-export default connect(
-    ({app}) => ({searchIsActive: app.searchIsActive}),
-    dispatch => bindActionCreators({askForGeoLocation}, dispatch)
-)(App);
+export default withRouter(connect(
+    ({app, routing}) => ({
+        searchIsActive: app.searchIsActive,
+        pathname: routing.location.pathname
+    }),
+    dispatch => bindActionCreators({askForGeoLocation, getConnectedUserToken}, dispatch)
+)(App));