+++ src/containers/Filters.js
@@ -1,5 +1,6 @@
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
+import { push } from 'react-router-redux';
 
 import Filters from '../components/Filters';
 import { searchUsers } from '../ducks/users';
@@ -8,12 +9,13 @@ import { updateGeoLocation } from '../ducks/geoLocation';
 import { activeSearch, disableSearch } from '../ducks/app';
 
 export default connect(
-    ({geoLocation, languages, app}) => ({geoLocation, languages}),
+    ({geoLocation, languages, routing}) => ({geoLocation, languages, pathname: routing.location.pathname}),
     dispatch => bindActionCreators({
       searchUsers,
       updateGeoLocation,
       updateLanguages,
       activeSearch,
-      disableSearch
+      disableSearch,
+      push
     }, dispatch)
 )(Filters);